Modified in order to allow only to Administrador user makes changes.

+++ src/main/java/fransis/mpm/controller/VendedorController.java
@@ -13,6 +13,8 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.annotation.*;
 
+import java.security.Principal;
+
 /**
  * Created by francisco on 23/12/15.
  */
@@ -25,28 +27,41 @@ public class VendedorController {
 
 
     @RequestMapping("/vendedores")
-    public Page<Vendedor> obtenerLista(Pageable pageRequest){
-        return vendedorRepository.findAll(pageRequest);
+    public Page<Vendedor> obtenerLista(Pageable pageRequest, Principal principal){
+        if("Administrador".equals(principal.getName())){
+            return vendedorRepository.findAll(pageRequest);
+        }else{
+            //return vendedorRepository.findByUsername(principal.getName());
+            return null;
+        }
     }
 
     @RequestMapping("/vendedor/search")
-    public Page<Vendedor> filtrar(@RequestParam(value = "") String search, Pageable pageRequest){
-        return vendedorRepository.findByUsernameContainingIgnoreCaseOrNombreContainingIgnoreCaseOrApellidoContainingIgnoreCase(search,search,search,pageRequest);
+    public Page<Vendedor> filtrar(@RequestParam(value = "") String search, Pageable pageRequest, Principal principal){
+        if("Administrador".equals(principal.getName())){
+            return vendedorRepository.findByUsernameContainingIgnoreCaseOrNombreContainingIgnoreCaseOrApellidoContainingIgnoreCase(search,search,search,pageRequest);
+        }else{
+            return null;
+        }
     }
 
     @RequestMapping("/vendedor/agregar")
-    public Vendedor agregar(@RequestBody Vendedor vendedor){
-        return vendedorRepository.saveAndFlush(vendedor);
+    public Vendedor agregar(@RequestBody Vendedor vendedor, Principal principal){
+        if("Administrador".equals(principal.getName())){
+            return vendedorRepository.saveAndFlush(vendedor);
+        }else{
+            return null;
+        }
     }
 
     @RequestMapping("/vendedor/delete/{vendedorId}")
-    public void delete(@PathVariable Long vendedorId){
-
-        Vendedor vendedor = vendedorRepository.findOne(vendedorId);
-        if(!"administrador".equalsIgnoreCase(vendedor.getUsername())){
-            vendedorRepository.delete(vendedorId);
+    public void delete(@PathVariable Long vendedorId, Principal principal){
+        if("Administrador".equals(principal.getName())){
+            Vendedor vendedor = vendedorRepository.findOne(vendedorId);
+            if(!"administrador".equalsIgnoreCase(vendedor.getUsername())){
+                vendedorRepository.delete(vendedorId);
+            }
         }
-
     }
 
 }
